refactor(metrics): Methods reorganized for maintainability
