+++ src/Classes/Requests/MetricRequests.php
@@ -11,11 +11,10 @@ use Anibalealvarezs\KlaviyoApi\Enums\AggregatedMeasurement;
 use Carbon\Carbon;
 use Classes\Conversions\GoogleSearchConsoleConvert;
 use Classes\Conversions\KlaviyoConvert;
-use Classes\KeyGenerator;
+use Classes\MetricsProcessor;
 use Classes\Overrides\GoogleApi\SearchConsoleApi\SearchConsoleApi;
 use Classes\Overrides\KlaviyoApi\KlaviyoApi;
 use DateTime;
-use DateTimeImmutable;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
@@ -35,6 +34,7 @@ use Enums\Channel;
 use Enums\Country as CountryEnum;
 use Enums\Device as DeviceEnum;
 use GuzzleHttp\Exception\GuzzleException;
+use Helpers\GoogleSearchConsoleHelpers;
 use Helpers\Helpers;
 use Monolog\Handler\StreamHandler;
 use Monolog\Level;
@@ -274,7 +274,7 @@ class MetricRequests
             // Custom filter for dimensions disabled for GSC given the strict structure. Config dimensions used instead
             $batchSize = 100;
 
-            $logger->info("Initialized repositories, dimensions=" . implode(',', GoogleSearchConsoleConvert::$allDimensions) . ", metricNames=" . json_encode($metricNames) . ", batchSize=$batchSize");
+            $logger->info("Initialized repositories, dimensions=" . implode(',', GoogleSearchConsoleHelpers::$allDimensions) . ", metricNames=" . json_encode($metricNames) . ", batchSize=$batchSize");
             $logger->warning("Note: 'searchAppearance' is not included in dimensions due to GSC API restrictions; defaulting to 'WEB' in ChanneledMetricDimension");
 
             // Load countries and create a map
@@ -332,8 +332,6 @@ class MetricRequests
                     $manager,
                     $channeledMetricRepository,
                     $pageRepository,
-                    $countryRepository,
-                    $deviceRepository,
                     $metricNames,
                     $filters,
                     $logger,
@@ -411,21 +409,7 @@ class MetricRequests
      */
     private static function validateGoogleConfig(LoggerInterface $logger): array
     {
-        $config = Helpers::getChannelsConfig()['google'] ?? null;
-        if (!$config) {
-            $logger->error("Missing 'google' configuration in channels config");
-            throw new Exception("Missing 'google' configuration in channels config");
-        }
-        $scConfig = Helpers::getChannelsConfig()['google_search_console'] ?? null;
-        if (!$scConfig) {
-            $logger->error("Missing 'google_search_console' configuration in channels config");
-            throw new Exception("Missing 'google_search_console' configuration in channels config");
-        }
-        $logger->info("Loaded GSC config: sites=" . count($scConfig['sites']));
-        return [
-            'google' => $config,
-            'google_search_console' => $scConfig,
-        ];
+        return GoogleSearchConsoleHelpers::validateGoogleConfig($logger);
     }
 
     /**
@@ -477,8 +461,6 @@ class MetricRequests
      * @param EntityManager $manager
      * @param EntityRepository $channeledMetricRepository
      * @param EntityRepository $pageRepository
-     * @param EntityRepository $countryRepository
-     * @param EntityRepository $deviceRepository
      * @param array $metricNames
      * @param object|null $filters
      * @param LoggerInterface $logger
@@ -502,8 +484,6 @@ class MetricRequests
         EntityManager $manager,
         EntityRepository $channeledMetricRepository,
         EntityRepository $pageRepository,
-        EntityRepository $countryRepository,
-        EntityRepository $deviceRepository,
         array $metricNames,
         ?object $filters,
         LoggerInterface $logger,
@@ -540,11 +520,6 @@ class MetricRequests
         list($from, $to) = self::determineDateRange($startDate, $lastChanneledMetric, $resume, $endDate, $logger);
 
         // Initialize daily processing
-        $entitiesToInvalidate = ['metric' => [], 'channeledMetric' => [], 'query' => []];
-        $queryCache = [];
-        $metricCache = [];
-        $channeledMetricCache = [];
-        $dimensionCache = [];
         $startTime = microtime(true);
         $siteMetrics = 0;
         $siteRows = 0;
@@ -559,13 +534,6 @@ class MetricRequests
                 $manager,
                 $pageEntity,
                 $metricNames,
-                $entitiesToInvalidate,
-                $queryCache,
-                $metricCache,
-                $channeledMetricCache,
-                $dimensionCache,
-                $countryRepository,
-                $deviceRepository,
                 $targetKeywords,
                 $targetCountries,
                 $dimensionFilterGroups,
@@ -600,13 +568,6 @@ class MetricRequests
      * @param EntityManager $manager
      * @param Page $pageEntity
      * @param array $metricNames
-     * @param array &$entitiesToInvalidate
-     * @param array &$queryCache
-     * @param array &$metricCache
-     * @param array &$channeledMetricCache
-     * @param array &$dimensionCache
-     * @param EntityRepository $countryRepository
-     * @param EntityRepository $deviceRepository
      * @param array $targetKeywords
      * @param array $targetCountries
      * @param array $dimensionFilterGroups
@@ -627,13 +588,6 @@ class MetricRequests
         EntityManager $manager,
         Page $pageEntity,
         array $metricNames,
-        array &$entitiesToInvalidate,
-        array &$queryCache,
-        array &$metricCache,
-        array &$channeledMetricCache,
-        array &$dimensionCache,
-        EntityRepository $countryRepository,
-        EntityRepository $deviceRepository,
         array $targetKeywords,
         array $targetCountries,
         array $dimensionFilterGroups,
@@ -648,7 +602,6 @@ class MetricRequests
         // $logger->info("Processing GSC data for site $siteUrl, date $dayStr");
 
         // Initialize counters
-        $loopCount = 0;
         $totalMetrics = 0;
         $totalRows = 0;
         $totalDuplicates = 0;
@@ -657,10 +610,10 @@ class MetricRequests
         $subsetRows = [];
 
         try {
-            $dimensionsSubsets = self::getAllSubsets(GoogleSearchConsoleConvert::$optionalDimensions);
+            $dimensionsSubsets = Helpers::getAllSubsets(GoogleSearchConsoleHelpers::$optionalDimensions);
             // $dimensionsSubsets = [GoogleSearchConsoleConvert::$optionalDimensions];
             foreach ($dimensionsSubsets as $dimensionsSubset) {
-                $actualDimensionsSubset = array_merge(array_diff(GoogleSearchConsoleConvert::$allDimensions, GoogleSearchConsoleConvert::$optionalDimensions), $dimensionsSubset);
+                $actualDimensionsSubset = array_merge(array_diff(GoogleSearchConsoleHelpers::$allDimensions, GoogleSearchConsoleHelpers::$optionalDimensions), $dimensionsSubset);
                 $rows = $api->getAllSearchQueryResults(
                     siteUrl: $siteUrl,
                     startDate: $dayStr,
@@ -677,7 +630,7 @@ class MetricRequests
 
             foreach($subsetRows as $rows) {
                 foreach ($rows as $row) {
-                    foreach($row as $key => $element) {
+                    foreach($row as $element) {
                         if (is_string($element)) {
                             continue;
                         }
@@ -687,31 +640,7 @@ class MetricRequests
                 }
             }
 
-            $childrenSums = self::computeChildrenSum($allRows);
-
-            /* $sums = [];
-            foreach ($allRows as $i => $row) {
-                $sums[] = "Record $i has children sum = " . (is_array($childrenSums[$i]) ? json_encode($childrenSums[$i]) : $childrenSums[$i]);
-            } */
-
-            $differences = self::calculateDifferences($allRows, $childrenSums);
-
-            $allocatedDifferences = self::allocatePositiveDifferences($differences, GoogleSearchConsoleConvert::$allDimensions);
-
-            $allocateFinalDifference = self::addGlobalRemainderSynthetic($allocatedDifferences, GoogleSearchConsoleConvert::$allDimensions);
-
-            $negativeDifferencesProcessed = self::flagOrScaleNegativeDifferences($allocateFinalDifference, true);
-
-            $scaleAdjusted = self::adjustScaledPositions($negativeDifferencesProcessed);
-
-            $finalRecords = array_values(array_filter($scaleAdjusted, function($record) {
-                // Keep if:
-                // - full subset (all 5 dims)
-                // - or synthetic record
-                return (count($record['subset']) === 5) || (!empty($record['synthetic']));
-            }));
-
-            $finalRecords = GoogleSearchConsoleConvert::fillWithNullsAndFilter($finalRecords, $targetKeywords, $targetCountries);
+            $finalRecords = GoogleSearchConsoleHelpers::getFinalRecords($allRows, $targetKeywords, $targetCountries);
 
             // Helpers::dumpDebugJson($finalRecords);
 
@@ -738,7 +667,7 @@ class MetricRequests
                 $manager->getConnection()->beginTransaction();
 
                 // Map metrics
-                $metricMap = self::processMetrics(
+                $metricMap = MetricsProcessor::processMetrics(
                     $allMetrics,
                     $manager,
                     $countryMap,
@@ -747,7 +676,7 @@ class MetricRequests
                 );
 
                 // Map channeled metrics
-                $channeledMetricMap = self::processChanneledMetrics(
+                $channeledMetricMap = MetricsProcessor::processChanneledMetrics(
                     $allMetrics,
                     $manager,
                     $metricMap,
@@ -755,7 +684,7 @@ class MetricRequests
                 );
 
                 // Map dimensions
-                self::processChanneledMetricDimensions(
+                MetricsProcessor::processChanneledMetricDimensions(
                     $allMetrics,
                     $manager,
                     $metricMap,
@@ -781,335 +710,6 @@ class MetricRequests
             $logger->error("Error during GSC API query for site $siteUrl, date $dayStr: " . $e->getMessage() . ", trace: " . $e->getTraceAsString());
             throw $e;
         }
-    }
-
-    protected static function computeChildrenSum(array $records): array {
-        $n = count($records);
-
-        // Initialize sums array with zeros for each metric
-        $childrenSums = array_fill(0, $n, ['impressions' => 0, 'clicks' => 0]);
-
-        // Sum children metrics
-        for ($i = 0; $i < $n; $i++) {
-            $parentSubset = $records[$i]['subset'];
-            $parentDims = $records[$i]['keys'];
-
-            for ($j = 0; $j < $n; $j++) {
-                if ($i === $j) continue;
-
-                $childSubset = $records[$j]['subset'];
-                $childDims = $records[$j]['keys'];
-
-                if (self::isParentOf($parentSubset, $parentDims, $childSubset, $childDims)) {
-                    $childrenSums[$i]['impressions'] += $records[$j]['impressions'] ?? 0;
-                    $childrenSums[$i]['clicks'] += $records[$j]['clicks'] ?? 0;
-                }
-            }
-        }
-
-        return $childrenSums;
-    }
-
-    protected static function calculateDifferences(array $records, array $childrenSums): array {
-        foreach ($records as $index => &$record) {
-            $ownImpressions = $record['impressions'] ?? 0;
-            $ownClicks = $record['clicks'] ?? 0;
-
-            $childrenImpressions = $childrenSums[$index]['impressions'] ?? 0;
-            $childrenClicks = $childrenSums[$index]['clicks'] ?? 0;
-
-            $record['impressions_difference'] = $ownImpressions - $childrenImpressions;
-            $record['clicks_difference'] = $ownClicks - $childrenClicks;
-        }
-        unset($record); // prevent accidental reference reuse
-        return $records;
-    }
-
-    protected static function allocatePositiveDifferences(array $records, array $dimensionNames, string $missingLabel = 'unknown'): array {
-        $extendedRecords = $records;
-
-        foreach ($records as $record) {
-            $impressionDiff = $record['impressions_difference'] ?? 0;
-            $clicksDiff = $record['clicks_difference'] ?? 0;
-
-            if ($impressionDiff > 0 || $clicksDiff > 0) {
-                $newKeys = $record['keys'];
-                $subset = $record['subset'];
-
-                // Find the first missing dimension
-                $missingDimension = null;
-                foreach ($dimensionNames as $dim) {
-                    if (!in_array($dim, $subset)) {
-                        $missingDimension = $dim;
-                        break;
-                    }
-                }
-
-                if ($missingDimension !== null) {
-                    $newKeys[] = $missingLabel;
-                    $newSubset = [...$subset, $missingDimension];
-
-                    // Create synthetic record
-                    $syntheticRecord = [
-                        'keys' => $newKeys,
-                        'clicks' => $clicksDiff,
-                        'impressions' => $impressionDiff,
-                        'ctr' => ($impressionDiff > 0) ? $clicksDiff / $impressionDiff : 0,
-                        'position' => null,
-                        'subset' => $newSubset,
-                        'impressions_difference' => 0,
-                        'clicks_difference' => 0,
-                        'synthetic' => true,
-                    ];
-
-                    $extendedRecords[] = $syntheticRecord;
-                }
-            }
-        }
-
-        return $extendedRecords;
-    }
-
-    protected static function addGlobalRemainderSynthetic(array $records, array $dimensionNames, array $parentSubset = ['date', 'page']): array
-    {
-        $extendedRecords = $records;
-
-        $allImpressions = 0;
-        $allClicks = 0;
-        $allPositionWeightedSum = 0;
-        $allPositionCount = 0;
-
-        $fiveDImpressions = 0;
-        $fiveDClicks = 0;
-        $fiveDPositionWeightedSum = 0;
-        $fiveDPositionCount = 0;
-
-        $partialImpressions = 0;
-        $partialClicks = 0;
-        $partialPositionWeightedSum = 0;
-        $partialPositionCount = 0;
-
-        // Step 1: Sum metrics by category
-        foreach ($records as $rec) {
-            $subset = $rec['subset'] ?? [];
-            $impr = $rec['impressions'] ?? 0;
-            $clicks = $rec['clicks'] ?? 0;
-            $pos = $rec['position'] ?? null;
-            $posWeighted = ($pos !== null) ? ($pos * $impr) : 0;
-
-            // Parent subset totals ("All")
-            if (array_values($subset) == $parentSubset) {
-                $allImpressions += $impr;
-                $allClicks += $clicks;
-                if ($pos !== null) {
-                    $allPositionWeightedSum += $posWeighted;
-                    $allPositionCount += $impr;
-                }
-            }
-
-            // Fully attributed 5D records (exclude synthetic)
-            if (count($subset) === count($dimensionNames) && empty($rec['synthetic'])) {
-                $fiveDImpressions += $impr;
-                $fiveDClicks += $clicks;
-                if ($pos !== null) {
-                    $fiveDPositionWeightedSum += $posWeighted;
-                    $fiveDPositionCount += $impr;
-                }
-            }
-
-            // Partial synthetic records
-            if (!empty($rec['synthetic'])) {
-                $partialImpressions += $impr;
-                $partialClicks += $clicks;
-                if ($pos !== null) {
-                    $partialPositionWeightedSum += $posWeighted;
-                    $partialPositionCount += $impr;
-                }
-            }
-        }
-
-        // Step 2: Calculate remaining values
-        $remainingImpressions = $allImpressions - $fiveDImpressions - $partialImpressions;
-        $remainingClicks = $allClicks - $fiveDClicks - $partialClicks;
-
-        // Weighted position average for remaining
-        // Compute weighted position of "All" minus weighted positions already accounted for
-        $allPositionAvg = ($allPositionCount > 0) ? ($allPositionWeightedSum / $allPositionCount) : null;
-
-        // For position, approximate remaining weighted sum and count
-        $remainingPositionWeightedSum = $allPositionWeightedSum - $fiveDPositionWeightedSum - $partialPositionWeightedSum;
-        $remainingPositionCount = $allPositionCount - $fiveDPositionCount - $partialPositionCount;
-
-        $remainingPosition = ($remainingPositionCount > 0) ? ($remainingPositionWeightedSum / $remainingPositionCount) : $allPositionAvg;
-
-        // CTR = clicks / impressions (avoid division by zero)
-        $remainingCtr = ($remainingImpressions > 0) ? ($remainingClicks / $remainingImpressions) : 0;
-
-        if ($remainingImpressions > 0 || $remainingClicks > 0) {
-            // Compose keys for parent subset, fill missing with $missingLabel
-            $keys = [];
-            foreach ($dimensionNames as $dim) {
-                $missingLabel = match($dim) {
-                    'country' => 'UNK',
-                    default => 'unknown',
-                };
-                if (in_array($dim, $parentSubset)) {
-                    // Find record with this subset and take its key for that dim, or fallback to missingLabel
-                    $foundKey = $missingLabel;
-                    foreach ($records as $rec) {
-                        if (($rec['subset'] ?? []) === $parentSubset) {
-                            $index = array_search($dim, $parentSubset);
-                            $foundKey = $rec['keys'][$index] ?? $missingLabel;
-                            break;
-                        }
-                    }
-                    $keys[] = $foundKey;
-                } else {
-                    $keys[] = $missingLabel;
-                }
-            }
-
-            $syntheticRecord = [
-                'keys' => $keys,
-                'subset' => $dimensionNames,
-                'impressions' => $remainingImpressions,
-                'clicks' => $remainingClicks,
-                'ctr' => $remainingCtr,
-                'position' => $remainingPosition,
-                'synthetic' => true,
-                'note' => 'final synthetic to reconcile unmatched parent metrics',
-                'impressions_difference' => 0,
-                'clicks_difference' => 0,
-            ];
-
-            $extendedRecords[] = $syntheticRecord;
-        }
-
-        return $extendedRecords;
-    }
-
-    protected static function flagOrScaleNegativeDifferences(array $records, bool $scaleNegative = false): array {
-        foreach ($records as &$record) {
-
-            $impressionDiff = $record['impressions_difference'] ?? 0;
-            $clicksDiff = $record['clicks_difference'] ?? 0;
-
-            $childrenImpressions = $record['children_sum']['impressions'] ?? 0;
-            $childrenClicks = $record['children_sum']['clicks'] ?? 0;
-
-            $impressions = $record['impressions'] ?? 0;
-            $clicks = $record['clicks'] ?? 0;
-
-            // If either metric has a negative difference, treat the record
-            if ($impressionDiff < 0 || $clicksDiff < 0) {
-                if ($scaleNegative) {
-                    $scaleFactorImpr = $childrenImpressions > 0 ? $impressions / $childrenImpressions : 0;
-                    $scaleFactorClicks = $childrenClicks > 0 ? $clicks / $childrenClicks : 0;
-
-                    $record['original_impressions'] = $impressions;
-                    $record['original_clicks'] = $clicks;
-                    $record['original_differences'] = [
-                        'impressions' => $impressionDiff,
-                        'clicks' => $clicksDiff
-                    ];
-
-                    $record['impressions'] = round($childrenImpressions * $scaleFactorImpr);
-                    $record['clicks'] = round($childrenClicks * $scaleFactorClicks);
-                    $record['scaled'] = true;
-                    $record['note'] = 'scaled down to match parent metrics';
-
-                    // Recalculate CTR
-                    $record['ctr'] = $record['impressions'] > 0 ? $record['clicks'] / $record['impressions'] : 0;
-                } else {
-                    $record['flagged'] = true;
-                    $record['note'] = 'exceeds parent; likely misattributed';
-                }
-            }
-        }
-
-        return $records;
-    }
-
-    protected static function adjustScaledPositions(array $records): array {
-        $n = count($records);
-
-        // Helper: same as before
-
-
-        for ($i = 0; $i < $n; $i++) {
-            if (!($records[$i]['scaled'] ?? false)) continue;
-
-            $parentSubset = $records[$i]['subset'];
-            $parentDims = $records[$i]['keys'];
-
-            $weightedSum = 0;
-            $totalImpressions = 0;
-
-            for ($j = 0; $j < $n; $j++) {
-                if ($i === $j) continue;
-
-                $childSubset = $records[$j]['subset'];
-                $childDims = $records[$j]['keys'];
-
-                if (self::isParentOf($parentSubset, $parentDims, $childSubset, $childDims)) {
-                    $impressions = $records[$j]['impressions'] ?? 0;
-                    $position = $records[$j]['position'] ?? null;
-
-                    if ($impressions > 0 && $position !== null) {
-                        $weightedSum += $impressions * $position;
-                        $totalImpressions += $impressions;
-                    }
-                }
-            }
-
-            if ($totalImpressions > 0) {
-                $records[$i]['original_position'] = $records[$i]['position'] ?? null;
-                $records[$i]['position'] = round($weightedSum / $totalImpressions, 2);
-            }
-        }
-
-        return $records;
-    }
-
-    protected static function isParentOf(array $parentSubset, array $parentDims, array $childSubset, array $childDims): bool {
-        if (count($childSubset) <= count($parentSubset)) {
-            return false;
-        }
-        $childSubsetIndex = array_flip($childSubset);
-        $parentIndexInChild = [];
-
-        foreach ($parentSubset as $dimName) {
-            if (!isset($childSubsetIndex[$dimName])) {
-                return false;
-            }
-            $parentIndexInChild[] = $childSubsetIndex[$dimName];
-        }
-
-        $prevIdx = -1;
-        foreach ($parentIndexInChild as $i => $childIdx) {
-            if ($childIdx <= $prevIdx) {
-                return false;
-            }
-            $prevIdx = $childIdx;
-            if ($parentDims[$i] !== $childDims[$childIdx]) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public static function getAllSubsets($elements): array
-    {
-        $subsets = [[]]; // Incluye el conjunto vacío
-        foreach ($elements as $element) {
-            $newSubsets = [];
-            foreach ($subsets as $subset) {
-                $newSubsets[] = $subset; // Copia el subconjunto existente
-                $newSubsets[] = array_merge($subset, [$element]); // Añade el elemento
-            }
-            $subsets = $newSubsets;
-        }
-        return $subsets;
     }
 
     /**
@@ -2301,1020 +1901,5 @@ class MetricRequests
             ];
         }
         return $dimensionFilterGroups;
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processQueries(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        // Extract queries from metrics
-        $queries = array_map(function ($metric) {
-            return $metric->query;
-        }, $metrics->toArray());
-
-        // Remove duplicates
-        $uniqueQueries = array_unique($queries);
-
-        // Batch select queries from list
-        $selectParams = array_values($uniqueQueries);
-        $selectPlaceholders = implode(', ', array_fill(0, count($selectParams), '?'));
-
-        $sql = "SELECT id, query
-                FROM queries
-                WHERE query IN ($selectPlaceholders)";
-        try {
-            $existingQueries = $manager->getConnection()
-                ->executeQuery($sql, $selectParams)
-                ->fetchAllAssociative();
-        } catch (\Doctrine\DBAL\Exception $e) {
-            throw new RuntimeException("Failed to fetch existing queries: " . $e->getMessage(), 0, $e);
-        }
-
-        // Map queries to their IDs and create a map for quick access
-        $map = [];
-        foreach ($existingQueries as $query) {
-            $map[$query['query']] = $query['id'];
-        }
-
-        // Get the list of queries that need to be inserted
-        $queriesToInsert = array_diff($uniqueQueries, array_keys($map));
-
-        // Bulk Insert queries that are not in the database with a single statement, letting SQL handle the auto-increment
-        if (!empty($queriesToInsert)) {
-            $insertPlaceholders = implode(', ', array_fill(0, count($queriesToInsert), '(?)'));
-            try {
-                $manager->getConnection()->executeStatement(
-                    "INSERT INTO queries (query) VALUES $insertPlaceholders",
-                    array_values($queriesToInsert)
-                );
-            } catch (\Doctrine\DBAL\Exception $e) {
-                throw new RuntimeException("Failed to insert queries: " . $e->getMessage(), 0, $e);
-            }
-        }
-
-        $allQueries = array_merge(array_keys($map), array_values($queriesToInsert));
-        $selectParams = array_values($allQueries);
-        $selectPlaceholders = implode(', ', array_fill(0, count($selectParams), '?'));
-
-        $sql = "SELECT id, query
-                FROM queries
-                WHERE query IN ($selectPlaceholders)";
-        try {
-            $finalQueries = $manager->getConnection()
-                ->executeQuery($sql, $selectParams)
-                ->fetchAllAssociative();
-        } catch (\Doctrine\DBAL\Exception $e) {
-            throw new RuntimeException("Failed to fetch existing queries: " . $e->getMessage(), 0, $e);
-        }
-        foreach ($finalQueries as $query) {
-            $map[$query['query']] = $query['id'];
-        }
-
-        return [
-            'map' => $map,
-            'mapReverse' => array_flip($map),
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processCampaigns(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processChanneledCampaigns(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processChanneledAdGroups(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processChanneledAds(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processPosts(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processProducts(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processCustomers(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @return array
-     */
-    protected static function processOrders(ArrayCollection $metrics, EntityManager $manager): array
-    {
-        return [
-            'map' => [],
-            'mapReverse' => [],
-        ];
-    }
-
-    /**
-     * @throws \Doctrine\DBAL\Exception
-     */
-    protected static function processMetrics(
-        ArrayCollection $metrics,
-        EntityManager $manager,
-        array $countryMap,
-        array $deviceMap,
-        array $pageMap,
-    ): array {
-
-        // Map queries
-        $queryMap = self::processQueries(
-            $metrics,
-            $manager,
-        );
-
-        // Map campaigns
-        $campaignMap = self::processCampaigns(
-            $metrics,
-            $manager,
-        );
-
-        // Map channeled campaigns
-        $channeledCampaignMap = self::processChanneledCampaigns(
-            $metrics,
-            $manager,
-        );
-
-        // Map channeled campaigns
-        $channeledAdGroupMap = self::processChanneledAdGroups(
-            $metrics,
-            $manager,
-        );
-
-        // Map channeled ads
-        $channeledAdMap = self::processChanneledAds(
-            $metrics,
-            $manager,
-        );
-
-        // Map posts
-        $postMap = self::processPosts(
-            $metrics,
-            $manager,
-        );
-
-        // Map products
-        $productMap = self::processProducts(
-            $metrics,
-            $manager,
-        );
-
-        // Map customers
-        $customerMap = self::processCustomers(
-            $metrics,
-            $manager,
-        );
-
-        // Map orders
-        $orderMap = self::processOrders(
-            $metrics,
-            $manager,
-        );
-
-        // Extract metrics from metrics
-        $uniqueMetrics = [];
-        foreach ($metrics->toArray() as $metric) {
-            $metricKey = KeyGenerator::generateMetricKey(
-                channel: $metric->channel,
-                name: $metric->name,
-                period: $metric->period,
-                metricDate: $metric->metricDate instanceof DateTime ? $metric->metricDate->format('Y-m-d') : $metric->metricDate,
-                campaign: isset($metric->campaign) ? $metric->campaign->getCampaignId() : null,
-                channeledCampaign: isset($metric->channeledCampaign) ? $metric->channeledCampaign->getPlatformId() : null,
-                channeledAdGroup: isset($metric->channeledAdGroup) ? $metric->channeledAdGroup->getPlatformId() : null,
-                channeledAd: isset($metric->channeledAd) ? $metric->channeledAd->getPlatformId() : null,
-                page: isset($metric->page) ? $metric->page->getUrl() : null,
-                query: $metric->query ?? null,
-                post: isset($metric->post) ? $metric->post->getPostId() : null,
-                product: isset($metric->product) ? $metric->product->getProductId() : null,
-                customer: isset($metric->customer) ? $metric->customer->getEmail() : null,
-                order: isset($metric->order) ? $metric->order->getOrderId() : null,
-                country: $metric->countryCode ?? null,
-                device: $metric->deviceType ?? null,
-            );
-
-            $uniqueMetrics[$metricKey] = [
-                'channel' => $metric->channel,
-                'name' => $metric->name,
-                'period' => $metric->period,
-                'metricDate' => $metric->metricDate instanceof DateTime ? $metric->metricDate->format('Y-m-d') : $metric->metricDate,
-                'campaign_id' => isset($metric->campaign) ? $campaignMap['map'][$metric->campaign->getCampaignId()] ?? null : null,
-                'channeledCampaign_id' => isset($metric->channeledCampaign) ? $channeledCampaignMap['map'][$metric->channeledCampaign->getPlatformId()] ?? null : null,
-                'channeledAdGroup_id' => isset($metric->channeledAdGroup) ? $channeledAdGroupMap['map'][$metric->channeledAdGroup->getPlatformId()] ?? null : null,
-                'channeledAd_id' => isset($metric->channeledAd) ? $channeledAdMap['map'][$metric->channeledAd->getPlatformId()] ?? null : null,
-                'query_id' => isset($metric->query) ? $queryMap['map'][$metric->query] : null,
-                'page_id' => isset($metric->page) ? $metric->page->getId() : null,
-                'post_id' => isset($metric->post) ? $postMap['map'][$metric->post->getPostId()] ?? null : null,
-                'product_id' => isset($metric->product) ? $productMap['map'][$metric->product->getProductId()] ?? null : null,
-                'customer_id' => isset($metric->customer) ? $customerMap['map'][$metric->customer->getEmail()] ?? null : null,
-                'order_id' => isset($metric->order) ? $orderMap['map'][$metric->order->getOrderId()] ?? null : null,
-                'country_id' => isset($metric->countryCode) ? $countryMap['map'][$metric->countryCode]->getId() : null,
-                'device_id' => isset($metric->deviceType) ? $deviceMap['map'][$metric->deviceType]->getId() : null,
-                'value' => $metric->value,
-                'metadata' => $metric->metadata,
-                'key' => $metricKey,
-            ];
-        }
-
-        // Batch select metrics from list
-        $conditions = [];
-        $selectParams = [];
-
-        $fields = [
-            'channel', 'name', 'period', 'metricDate', 'campaign_id', 'channeledCampaign_id', 'channeledAdGroup_id',
-            'channeledAd_id', 'query_id', 'page_id', 'post_id', 'product_id', 'customer_id', 'order_id', 'country_id', 'device_id'
-        ];
-
-        foreach ($uniqueMetrics as $m) {
-            $subConditions = [];
-            foreach ($fields as $field) {
-                if (array_key_exists($field, $m) && $m[$field] === null) {
-                    $subConditions[] = "$field IS NULL";
-                } else {
-                    $subConditions[] = "$field = ?";
-                    $selectParams[] = $m[$field];
-                }
-            }
-            $conditions[] = '(' . implode(' AND ', $subConditions) . ')';
-        }
-
-        $sql = "SELECT id, " . implode(', ', $fields) . "
-                FROM metrics
-                WHERE " . (empty($conditions) ? '1=0' : implode(' OR ', $conditions));
-
-        // Map metrics to their IDs
-        $metricMap = self::getMetricMap(
-            manager: $manager,
-            sql: $sql,
-            params: $selectParams,
-            campaignMap: $campaignMap['mapReverse'],
-            channeledCampaignMap: $channeledCampaignMap['mapReverse'],
-            channeledAdGroupMap: $channeledAdGroupMap['mapReverse'],
-            channeledAdMap: $channeledAdMap['mapReverse'],
-            pageMap: $pageMap['mapReverse'],
-            queryMap: $queryMap['mapReverse'],
-            postMap: $postMap['mapReverse'],
-            productMap: $productMap['mapReverse'],
-            customerMap: $customerMap['mapReverse'],
-            orderMap: $orderMap['mapReverse'],
-            countryMap: $countryMap['mapReverse'],
-            deviceMap: $deviceMap['mapReverse']
-        );
-
-        // Get the list of metrics that need to be inserted
-        $metricsToInsert = [];
-        foreach ($uniqueMetrics as $key => $metric) {
-            if (!isset($metricMap[$key])) {
-                $metricsToInsert[] = [
-                    'channel' => $metric['channel'],
-                    'name' => $metric['name'],
-                    'period' => $metric['period'],
-                    'metricDate' => $metric['metricDate'],
-                    'campaign_id' => $metric['campaign_id'] ?? null,
-                    'channeledCampaign_id' => $metric['channeledCampaign_id'] ?? null,
-                    'channeledAdGroup_id' => $metric['channeledAdGroup_id'] ?? null,
-                    'channeledAd_id' => $metric['channeledAd_id'] ?? null,
-                    'query_id' => $metric['query_id'] ?? null,
-                    'page_id' => $metric['page_id'] ?? null,
-                    'post_id' => $metric['post_id'] ?? null,
-                    'product_id' => $metric['product_id'] ?? null,
-                    'customer_id' => $metric['customer_id'] ?? null,
-                    'order_id' => $metric['order_id'] ?? null,
-                    'country_id' => $metric['country_id'] ?? null,
-                    'device_id' => $metric['device_id'] ?? null,
-                    'value' => $metric['value'],
-                    'metadata' => json_encode($metric['metadata'] ?? []),
-                    'key' => $key,
-                ];
-            }
-        }
-
-        // Bulk Insert metrics
-        if (!empty($metricsToInsert)) {
-            $insertParams = [];
-            foreach ($metricsToInsert as $row) {
-                $insertParams[] = $row['channel'];
-                $insertParams[] = $row['name'];
-                $insertParams[] = $row['period'];
-                $insertParams[] = $row['metricDate'];
-                $insertParams[] = $row['campaign_id'] ?? null;
-                $insertParams[] = $row['channeledCampaign_id'] ?? null;
-                $insertParams[] = $row['channeledAdGroup_id'] ?? null;
-                $insertParams[] = $row['channeledAd_id'] ?? null;
-                $insertParams[] = $row['query_id'] ?? null;
-                $insertParams[] = $row['page_id'] ?? null;
-                $insertParams[] = $row['post_id'] ?? null;
-                $insertParams[] = $row['product_id'] ?? null;
-                $insertParams[] = $row['customer_id'] ?? null;
-                $insertParams[] = $row['order_id'] ?? null;
-                $insertParams[] = $row['country_id'] ?? null;
-                $insertParams[] = $row['device_id'] ?? null;
-                $insertParams[] = $row['value'];
-                $insertParams[] = $row['metadata'];
-            }
-            // $logger->info("Inserting " . count($metricsToInsert) . " new metrics");
-            $manager->getConnection()->executeStatement(
-                'INSERT INTO metrics (channel, name, period, metricDate, campaign_id, channeledCampaign_id, channeledAdGroup_id,
-                            channeledAd_id, query_id, page_id, post_id, product_id, customer_id, order_id, country_id, device_id, value, metadata)
-                     VALUES ' . implode(', ', array_fill(0, count($metricsToInsert), '(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)')),
-                        $insertParams
-                    );
-
-            // Re-fetch inserted metrics to get correct IDs
-            $reFetchParams = [];
-            $conditions = [];
-
-            $fields = [
-                'channel', 'name', 'period', 'metricDate',
-                'campaign_id', 'channeledCampaign_id', 'channeledAdGroup_id', 'channeledAd_id',
-                'query_id', 'page_id', 'post_id', 'product_id',
-                'customer_id', 'order_id', 'country_id', 'device_id'
-            ];
-
-            foreach ($metricsToInsert as $row) {
-                $subConditions = [];
-                foreach ($fields as $field) {
-                    if (!array_key_exists($field, $row) || $row[$field] === null) {
-                        $subConditions[] = "$field IS NULL";
-                    } else {
-                        $subConditions[] = "$field = ?";
-                        $reFetchParams[] = $row[$field];
-                    }
-                }
-                $conditions[] = '(' . implode(' AND ', $subConditions) . ')';
-            }
-
-            $reFetchSql = "SELECT id, " . implode(', ', $fields) . "
-                            FROM metrics
-                            WHERE " . (empty($conditions) ? '1=0' : implode(' OR ', $conditions));
-
-            $newMetrics = $manager->getConnection()->executeQuery($reFetchSql, $reFetchParams)->fetchAllAssociative();
-
-            foreach ($newMetrics as $metric) {
-                $metricKey = KeyGenerator::generateMetricKey(
-                    channel: $metric['channel'],
-                    name: $metric['name'],
-                    period: $metric['period'],
-                    metricDate: $metric['metricDate'],
-                    campaign: isset($metric['campaign_id']) ? $campaignMap['mapReverse'][$metric['campaign_id']] : null,
-                    channeledCampaign: isset($metric['channeledCampaign_id']) ? $channeledCampaignMap['mapReverse'][$metric['channeledCampaign_id']] : null,
-                    channeledAdGroup: isset($metric['channeledAdGroup_id']) ? $channeledAdGroupMap['mapReverse'][$metric['channeledAdGroup_id']] : null,
-                    channeledAd: isset($metric['channeledAd_id']) ? $channeledAdMap['mapReverse'][$metric['channeledAd_id']] : null,
-                    page: isset($metric['page_id']) ? $pageMap['mapReverse'][$metric['page_id']]->getUrl() : null,
-                    query: isset($metric['query_id']) ? $queryMap['mapReverse'][$metric['query_id']] : null,
-                    post: isset($metric['post_id']) ? $postMap['mapReverse'][$metric['post_id']] : null,
-                    product: isset($metric['product_id']) ? $productMap['mapReverse'][$metric['product_id']] : null,
-                    customer: isset($metric['customer_id']) ? $customerMap['mapReverse'][$metric['customer_id']] : null,
-                    order: isset($metric['order_id']) ? $orderMap['mapReverse'][$metric['order_id']] : null,
-                    country: isset($metric['country_id']) ? $countryMap['mapReverse'][$metric['country_id']]->getCode() : null,
-                    device: isset($metric['device_id']) ? $deviceMap['mapReverse'][$metric['device_id']]->getType() : null,
-                );
-                $metricMap[$metricKey] = (int)$metric['id'];
-                // $logger->info("Added metric to map: metricKey=$metricKey, metric_id={$metric['id']}");
-            }
-        }
-
-        return [
-            'map' => $metricMap,
-            'mapReverse' => array_flip($metricMap),
-        ];
-    }
-
-    /**
-     * @param ArrayCollection $metrics
-     * @param EntityManager $manager
-     * @param array $metricMap
-     * @param LoggerInterface $logger
-     * @return array
-     * @throws \Doctrine\DBAL\Exception
-     */
-    protected static function processChanneledMetrics(
-        ArrayCollection $metrics,
-        EntityManager $manager,
-        array $metricMap,
-        LoggerInterface $logger,
-    ): array {
-        // $logger->info("Processing " . count($metrics->toArray()) . " metrics in processChanneledMetrics");
-        // Extract channeled metrics from metrics
-        $uniqueChanneledMetrics = [];
-        foreach ($metrics->toArray() as $metric) {
-            $metricKey = KeyGenerator::generateMetricKey(
-                channel: $metric->channel,
-                name: $metric->name,
-                period: $metric->period,
-                metricDate: $metric->metricDate instanceof DateTime ? $metric->metricDate->format('Y-m-d') : $metric->metricDate,
-                campaign: isset($metric->campaign) ? $metric->campaign->getCampaignId() : null,
-                channeledCampaign: isset($metric->channeledCampaign) ? $metric->channeledCampaign->getPlatformId() : null,
-                channeledAdGroup: isset($metric->channeledAdGroup) ? $metric->channeledAdGroup->getPlatformId() : null,
-                channeledAd: isset($metric->channeledAd) ? $metric->channeledAd->getPlatformId() : null,
-                page: isset($metric->page) ? $metric->page->getUrl() : null,
-                query: $metric->query ?? null,
-                post: isset($metric->post) ? $metric->post->getPostId() : null,
-                product: isset($metric->product) ? $metric->product->getProductId() : null,
-                customer: isset($metric->customer) ? $metric->customer->getEmail() : null,
-                order: isset($metric->order) ? $metric->order->getOrderId() : null,
-                country: $metric->countryCode ?? null,
-                device: $metric->deviceType ?? null,
-            );
-
-            if (!isset($metricMap['map'][$metricKey])) {
-                $logger->warning("Skipping channeled metric due to missing metricKey: metricKey=$metricKey, query=$metric->query, page={$metric->page->getUrl()}");
-                continue;
-            }
-            if (empty($metric->platformId)) {
-                $logger->warning("Skipping channeled metric: platformId is empty, metricKey=$metricKey");
-                continue;
-            }
-
-            // $logger->info("ChanneledMetrics Inputs: metricKey=$metricKey, channel={$metric->channel}, platformId=" . ($metric->platformId ?? 'null') . ", metric_id={$metricMap['map'][$metricKey]}, platformCreatedAt=" . ($metric->platformCreatedAt->format('Y-m-d H:i:s')));
-            $channeledMetricKey = KeyGenerator::generateChanneledMetricKey(
-                channel: $metric->channel,
-                platformId: $metric->platformId,
-                metric: $metricMap['map'][$metricKey],
-                platformCreatedAt: Carbon::parse($metric->platformCreatedAt)->format('Y-m-d'),
-            );
-
-            $uniqueChanneledMetrics[$channeledMetricKey] = [
-                'channel' => $metric->channel,
-                'platformId' => $metric->platformId,
-                'metric_id' => $metricMap['map'][$metricKey],
-                'platformCreatedAt' => Carbon::parse($metric->platformCreatedAt)->format('Y-m-d'),
-                'data' => $metric->data ?? ['impressions' => 0, 'clicks' => 0, 'position_weighted' => 0, 'ctr' => 0],
-                'metricKey' => $metricKey,
-            ];
-            // $logger->info("Prepared channeled metric: channeledMetricKey=$channeledMetricKey, metricKey=$metricKey, metric_id={$metricMap['map'][$metricKey]}, platformId=$metric->platformId");
-        }
-
-        // Batch select channeled metrics from list
-        $conditions = [];
-        $selectParams = [];
-
-        foreach ($uniqueChanneledMetrics as $m) {
-            $conditions[] = '(channel = ? AND platformId = ? AND metric_id = ? AND platformCreatedAt = ?)';
-            $selectParams[] = $m['channel'];
-            $selectParams[] = $m['platformId'];
-            $selectParams[] = $m['metric_id'];
-            $selectParams[] = $m['platformCreatedAt'];
-        }
-
-        $sql = "SELECT id, channel, platformId, metric_id, platformCreatedAt, data
-            FROM channeled_metrics
-            WHERE " . (empty($conditions) ? '1=0' : implode(' OR ', $conditions));
-
-        // $logger->info("uniqueChanneledMetrics count: " . count($uniqueChanneledMetrics));
-        // Map metrics to their IDs
-        $channeledMetricMap = self::getChanneledMetricMap(
-            $manager,
-            $sql,
-            $selectParams,
-            $metricMap,
-        );
-        // $logger->info("channeledMetricMap count after re-fetch: " . count($channeledMetricMap));
-
-        // Get list of channeled metrics that need to be inserted and updated
-        $channeledMetricsToInsert = [];
-        $channeledMetricsToUpdate = [];
-        foreach ($uniqueChanneledMetrics as $key => $channeledMetric) {
-            if (!isset($channeledMetricMap[$key]) && !isset($channeledMetricsToInsert[$key])) {
-                $channeledMetricsToInsert[$key] = [
-                    'channel' => $channeledMetric['channel'],
-                    'platformId' => $channeledMetric['platformId'],
-                    'metric_id' => $channeledMetric['metric_id'],
-                    'platformCreatedAt' => $channeledMetric['platformCreatedAt'],
-                    'data' => json_encode($channeledMetric['data']),
-                ];
-                // $logger->info("Queuing channeled metric for insert: channeledMetricKey=$key, metric_id={$channeledMetric['metric_id']}, metricKey={$channeledMetric['metricKey']}, platformId={$channeledMetric['platformId']}");
-            } elseif (isset($channeledMetricsToInsert[$key])) {
-                // Update data
-                $data = json_decode($channeledMetricsToInsert[$key]['data'], true);
-                $newData = $channeledMetric['data'];
-
-                $data['impressions'] = max($data['impressions'] ?? 0, $newData['impressions'] ?? 0);
-                $data['clicks'] = max($data['clicks'] ?? 0, $newData['clicks'] ?? 0);
-                $data['position_weighted'] = max($data['position_weighted'] ?? 0, $newData['position_weighted'] ?? 0);
-                $data['ctr'] = max($data['ctr'] ?? 0, $newData['ctr'] ?? 0);
-
-                $channeledMetricsToInsert[$key]['data'] = json_encode($data);
-                // $logger->info("Updated queued channeled metric: channeledMetricKey=$key, metric_id={$channeledMetric['metric_id']}");
-            } else {
-                // Update existing
-                $data = json_decode($channeledMetricMap[$key]['data'], true);
-                $newData = $channeledMetric['data'];
-
-                $data['impressions'] = max($data['impressions'] ?? 0, $newData['impressions'] ?? 0);
-                $data['clicks'] = max($data['clicks'] ?? 0, $newData['clicks'] ?? 0);
-                $data['position_weighted'] = max($data['position_weighted'] ?? 0, $newData['position_weighted'] ?? 0);
-                $data['ctr'] = max($data['ctr'] ?? 0, $newData['ctr'] ?? 0);
-
-                $channeledMetricsToUpdate[$key] = [
-                    'id' => $channeledMetricMap[$key]['id'],
-                    'data' => json_encode($data, JSON_UNESCAPED_UNICODE),
-                ];
-                // $logger->info("Queuing channeled metric for update: channeledMetricKey=$key, metric_id={$channeledMetric['metric_id']}");
-            }
-        }
-
-        // Bulk Insert channeled metrics
-        if (!empty($channeledMetricsToInsert)) {
-            $insertPlaceholders = implode(', ', array_fill(0, count($channeledMetricsToInsert), '(?, ?, ?, ?, ?)'));
-            $insertParams = [];
-            foreach ($channeledMetricsToInsert as $row) {
-                $insertParams[] = $row['channel'];
-                $insertParams[] = $row['platformId'];
-                $insertParams[] = $row['metric_id'];
-                $insertParams[] = $row['platformCreatedAt'];
-                $insertParams[] = $row['data'];
-            }
-            // $logger->info("Inserting " . count($channeledMetricsToInsert) . " channeled metrics");
-            $manager->getConnection()->executeStatement(
-                'INSERT INTO channeled_metrics (channel, platformId, metric_id, platformCreatedAt, data)
-                    VALUES ' . $insertPlaceholders,
-                $insertParams
-            );
-
-            // Re-fetch channeled metrics to get their IDs
-            $selectParams = [];
-            $selectPlaceholders = [];
-            foreach ($channeledMetricsToInsert as $m) {
-                $selectPlaceholders[] = '(?, ?, ?, ?)';
-                $selectParams[] = $m['channel'];
-                $selectParams[] = $m['platformId'];
-                $selectParams[] = $m['metric_id'];
-                $selectParams[] = $m['platformCreatedAt'];
-            }
-            $sql = "SELECT id, channel, platformId, metric_id, platformCreatedAt, data
-                    FROM channeled_metrics
-                    WHERE (channel, platformId, metric_id, platformCreatedAt) IN (" . implode(', ', $selectPlaceholders) . ")";
-        }
-
-        // Bulk Update channeled metrics
-        if (!empty($channeledMetricsToUpdate)) {
-            $updateCases = [];
-            $updateParams = [];
-            $ids = [];
-            foreach ($channeledMetricsToUpdate as $update) {
-                $id = $update['id'];
-                $data = $update['data'];
-                $updateCases[] = "WHEN id = ? THEN ?";
-                $updateParams[] = $id;
-                $updateParams[] = $data;
-                $ids[] = $id;
-            }
-            $caseSql = implode("\n", $updateCases);
-            $idPlaceholders = implode(', ', array_fill(0, count($ids), '?'));
-            $whereParams = $ids;
-            $updateParams = array_merge($updateParams, $whereParams);
-            // $logger->info("Updating " . count($channeledMetricsToUpdate) . " channeled metrics");
-            $updateSql = "UPDATE channeled_metrics
-                      SET data = CASE
-                                     $caseSql
-                                     ELSE data
-                                 END
-                      WHERE id IN ($idPlaceholders)";
-            $manager->getConnection()->executeStatement($updateSql, $updateParams);
-        }
-
-        // Re-fetch channeled metrics
-        $channeledMetricMap = self::getChanneledMetricMap(
-            $manager,
-            $sql,
-            $selectParams,
-            $metricMap,
-        );
-        // $logger->info("uniqueChanneledMetrics count: " . count($uniqueChanneledMetrics));
-        // $logger->info("channeledMetricMap count after re-fetch: " . count($channeledMetricMap));
-
-        $channeledMetricMapFlipped = [];
-        foreach ($channeledMetricMap as $originalKey => $value) {
-            if (isset($value['id'])) {
-                $id = (string)$value['id'];
-                $channeledMetricMapFlipped[$id] = [
-                    'id' => $originalKey,
-                    'data' => $value['data'],
-                ];
-            }
-        }
-
-        return [
-            'channeledMetricMap' => $channeledMetricMap,
-            'channeledMetricMapReverse' => $channeledMetricMapFlipped,
-        ];
-    }
-
-    /**
-     * @throws NotSupported
-     * @throws ORMException
-     * @throws \Doctrine\DBAL\Exception
-     */
-    protected static function processChanneledMetricDimensions(
-        ArrayCollection $metrics,
-        EntityManager $manager,
-        array $metricMap,
-        array $channeledMetricMap,
-        LoggerInterface $logger
-    ): void {
-        // $logger->info("Processing " . count($metrics->toArray()) . " metrics in processChanneledMetricDimensions");
-        // Extract dimensions from metrics
-        $uniqueDimensions = [];
-        foreach ($metrics->toArray() as $metric) {
-            $dimensions = $metric->dimensions;
-            foreach ($dimensions as $dimension) {
-                $metricKey = KeyGenerator::generateMetricKey(
-                    channel: $metric->channel,
-                    name: $metric->name,
-                    period: $metric->period,
-                    metricDate: $metric->metricDate instanceof DateTime ? $metric->metricDate->format('Y-m-d') : $metric->metricDate,
-                    campaign: isset($metric->campaign) ? $metric->campaign->getCampaignId() : null,
-                    channeledCampaign: isset($metric->channeledCampaign) ? $metric->channeledCampaign->getPlatformId() : null,
-                    channeledAdGroup: isset($metric->channeledAdGroup) ? $metric->channeledAdGroup->getPlatformId() : null,
-                    channeledAd: isset($metric->channeledAd) ? $metric->channeledAd->getPlatformId() : null,
-                    page: isset($metric->page) ? $metric->page->getUrl() : null,
-                    query: $metric->query ?? null,
-                    post: isset($metric->post) ? $metric->post->getPostId() : null,
-                    product: isset($metric->product) ? $metric->product->getProductId() : null,
-                    customer: isset($metric->customer) ? $metric->customer->getEmail() : null,
-                    order: isset($metric->order) ? $metric->order->getOrderId() : null,
-                    country: $metric->countryCode ?? null,
-                    device: $metric->deviceType ?? null,
-                );
-
-                if (!isset($metricMap['map'][$metricKey])) {
-                    $logger->warning("Skipping dimension: metricKey=$metricKey not found in metricMap");
-                    continue;
-                }
-
-                // $logger->info("Dimensions Inputs: metricKey=$metricKey, channel={$metric->channel}, platformId=" . ($metric->platformId ?? 'null') . ", metric_id={$metricMap['map'][$metricKey]}, platformCreatedAt=" . ($metric->platformCreatedAt->format('Y-m-d H:i:s')));
-                $channeledMetricKey = KeyGenerator::generateChanneledMetricKey(
-                    channel: $metric->channel,
-                    platformId: $metric->platformId,
-                    metric: $metricMap['map'][$metricKey],
-                    platformCreatedAt: Carbon::parse($metric->platformCreatedAt)->format('Y-m-d'),
-                );
-
-                if (!isset($channeledMetricMap['channeledMetricMap'][$channeledMetricKey])) {
-                    $logger->error("ChanneledMetric not found for key=$channeledMetricKey, metricId=".$metricMap['map'][$metricKey]." metricKey=$metricKey, platformId=" . ($metric->platformId ?? 'null') . ", platformCreatedAt=" . Carbon::parse($metric->platformCreatedAt)->format('Y-m-d'));
-                    continue;
-                }
-
-                $dimensionKey = KeyGenerator::generateChanneledMetricDimensionKey(
-                    channeledMetric: $channeledMetricMap['channeledMetricMap'][$channeledMetricKey]['id'],
-                    dimensionKey: $dimension->dimensionKey,
-                    dimensionValue: $dimension->dimensionValue,
-                );
-
-                $uniqueDimensions[$dimensionKey] = [
-                    'dimensionKey' => $dimension->dimensionKey,
-                    'dimensionValue' => $dimension->dimensionValue,
-                    'channeledMetric_id' => $channeledMetricMap['channeledMetricMap'][$channeledMetricKey]['id'],
-                ];
-            }
-        }
-
-        // Batch select dimensions from list
-        $selectParams = [];
-        $selectPlaceholders = [];
-        $nullConditions = [];
-
-        foreach ($uniqueDimensions as $d) {
-            $channeledMetricId = $d['channeledMetric_id'];
-            $dimensionKey = $d['dimensionKey'];
-            $dimensionValue = $d['dimensionValue'];
-
-            if ($dimensionKey === null || $dimensionValue === null) {
-                // Use IS NULL-safe fallback condition for NULLs
-                $conditions = [];
-                $conditions[] = 'channeledMetric_id = ?';
-                $selectParams[] = $channeledMetricId;
-
-                $conditions[] = $dimensionKey === null ? 'dimensionKey IS NULL' : 'dimensionKey = ?';
-                if ($dimensionKey !== null) $selectParams[] = $dimensionKey;
-
-                $conditions[] = $dimensionValue === null ? 'dimensionValue IS NULL' : 'dimensionValue = ?';
-                if ($dimensionValue !== null) $selectParams[] = $dimensionValue;
-
-                $nullConditions[] = '(' . implode(' AND ', $conditions) . ')';
-            } else {
-                $selectParams[] = $channeledMetricId;
-                $selectParams[] = $dimensionKey;
-                $selectParams[] = $dimensionValue;
-                $selectPlaceholders[] = '(?, ?, ?)';
-            }
-        }
-
-        $whereParts = [];
-
-        if (!empty($selectPlaceholders)) {
-            $whereParts[] = '(channeledMetric_id, dimensionKey, dimensionValue) IN (' . implode(', ', $selectPlaceholders) . ')';
-        }
-
-        if (!empty($nullConditions)) {
-            $whereParts[] = implode(' OR ', $nullConditions);
-        }
-
-        $sql = "SELECT id, dimensionKey, dimensionValue, channeledMetric_id
-                FROM channeled_metric_dimensions";
-
-        if (!empty($whereParts)) {
-            $sql .= ' WHERE ' . implode(' OR ', $whereParts);
-        } else {
-            $sql .= ' WHERE 1=0'; // Safe fallback if no dimensions to match
-        }
-
-        // Map dimensions to their IDs
-        $dimensionMap = self::getDimensionMap(
-            $manager,
-            $sql,
-            $selectParams,
-        );
-
-        // Get list of dimensions that need to be inserted
-        $dimensionsToInsert = [];
-        foreach ($uniqueDimensions as $key => $dimension) {
-            if (!isset($dimensionMap[$key])) {
-                $dimensionsToInsert[] = [
-                    'channeledMetric_id' => $dimension['channeledMetric_id'],
-                    'dimensionKey' => $dimension['dimensionKey'],
-                    'dimensionValue' => $dimension['dimensionValue'],
-                ];
-            }
-        }
-
-        // Bulk Insert dimensions that are not in the database
-        if (!empty($dimensionsToInsert)) {
-            $insertPlaceholders = implode(', ', array_fill(0, count($dimensionsToInsert), '(?, ?, ?)'));
-            $insertParams = [];
-            foreach ($dimensionsToInsert as $dimensionToInsert) {
-                $insertParams[] = $dimensionToInsert['channeledMetric_id'];
-                $insertParams[] = $dimensionToInsert['dimensionKey'];
-                $insertParams[] = $dimensionToInsert['dimensionValue']; // null-safe
-            }
-
-            $manager->getConnection()->executeStatement(
-                "INSERT INTO channeled_metric_dimensions (channeledMetric_id, dimensionKey, dimensionValue)
-                    VALUES $insertPlaceholders",
-                $insertParams);
-
-            // Refetch dimensions to get their IDs
-            $selectParams = [];
-            $selectPlaceholders = [];
-            $nullConditions = [];
-
-            foreach ($dimensionsToInsert as $d) {
-                $channeledMetricId = $d['channeledMetric_id'];
-                $dimensionKey = $d['dimensionKey'];
-                $dimensionValue = $d['dimensionValue'];
-
-                if ($dimensionKey === null || $dimensionValue === null) {
-                    // Use IS NULL-safe fallback condition for NULLs
-                    $conditions = [];
-                    $conditions[] = 'channeledMetric_id = ?';
-                    $selectParams[] = $channeledMetricId;
-
-                    $conditions[] = $dimensionKey === null ? 'dimensionKey IS NULL' : 'dimensionKey = ?';
-                    if ($dimensionKey !== null) $selectParams[] = $dimensionKey;
-
-                    $conditions[] = $dimensionValue === null ? 'dimensionValue IS NULL' : 'dimensionValue = ?';
-                    if ($dimensionValue !== null) $selectParams[] = $dimensionValue;
-
-                    $nullConditions[] = '(' . implode(' AND ', $conditions) . ')';
-                } else {
-                    $selectParams[] = $channeledMetricId;
-                    $selectParams[] = $dimensionKey;
-                    $selectParams[] = $dimensionValue;
-                    $selectPlaceholders[] = '(?, ?, ?)';
-                }
-            }
-
-            $whereParts = [];
-
-            if (!empty($selectPlaceholders)) {
-                $whereParts[] = '(channeledMetric_id, dimensionKey, dimensionValue) IN (' . implode(', ', $selectPlaceholders) . ')';
-            }
-
-            if (!empty($nullConditions)) {
-                $whereParts[] = implode(' OR ', $nullConditions);
-            }
-
-            $sql = "SELECT id, dimensionKey, dimensionValue, channeledMetric_id
-                FROM channeled_metric_dimensions";
-
-            if (!empty($whereParts)) {
-                $sql .= ' WHERE ' . implode(' OR ', $whereParts);
-            } else {
-                $sql .= ' WHERE 1=0'; // Safe fallback if no dimensions to match
-            }
-        }
-
-        $manager->getConnection()
-            ->executeQuery($sql, $selectParams)
-            ->fetchAllAssociative();
-    }
-
-    /**
-     * @param EntityManager $manager
-     * @param string $sql
-     * @param array $params
-     * @param array $campaignMap
-     * @param array $channeledCampaignMap
-     * @param array $channeledAdGroupMap
-     * @param array $channeledAdMap
-     * @param array $pageMap
-     * @param array $queryMap
-     * @param array $postMap
-     * @param array $productMap
-     * @param array $customerMap
-     * @param array $orderMap
-     * @param array $countryMap
-     * @param array $deviceMap
-     * @return array
-     * @throws \Doctrine\DBAL\Exception
-     */
-    protected static function getMetricMap(
-        EntityManager $manager,
-        string $sql,
-        array $params,
-        array $campaignMap = [],
-        array $channeledCampaignMap = [],
-        array $channeledAdGroupMap = [],
-        array $channeledAdMap = [],
-        array $pageMap = [],
-        array $queryMap = [],
-        array $postMap = [],
-        array $productMap = [],
-        array $customerMap = [],
-        array $orderMap = [],
-        array $countryMap = [],
-        array $deviceMap = [],
-    ): array {
-        // Update the metric map with the newly inserted metrics
-        $existingMetrics = $manager->getConnection()
-            ->executeQuery($sql, $params)
-            ->fetchAllAssociative();
-
-        // Map metrics to their IDs and create a map for quick access
-        $metricMap = [];
-        foreach ($existingMetrics as $metric) {
-            $metricKey = KeyGenerator::generateMetricKey(
-                channel: $metric['channel'],
-                name: $metric['name'],
-                period: $metric['period'],
-                metricDate: $metric['metricDate'],
-                campaign: isset($metric['campaign_id']) ? $campaignMap[$metric['campaign_id']]->getCampaignId() : null,
-                channeledCampaign: isset($metric['channeledCampaign_id']) ? $channeledCampaignMap[$metric['channeledCampaign_id']]->getPlatformId() : null,
-                channeledAdGroup: isset($metric['channeledAdGroup_id']) ? $channeledAdGroupMap[$metric['channeledAdGroup_id']]->getPlatformId() : null,
-                channeledAd: isset($metric['channeledAd_id']) ? $channeledAdMap[$metric['channeledAd_id']]->getPlatformId() : null,
-                page: isset($metric['page_id']) ? $pageMap[$metric['page_id']]->getUrl() : null,
-                query: isset($metric['query_id']) ? $queryMap[$metric['query_id']] : null,
-                post: isset($metric['post_id']) ? $postMap[$metric['post_id']]->getPostId() : null,
-                product: isset($metric['product_id']) ? $productMap[$metric['product_id']]->getProductId() : null,
-                customer: isset($metric['customer_id']) ? $customerMap[$metric['customer_id']]->getEmail() : null,
-                order: isset($metric['order_id']) ? $orderMap[$metric['order_id']]->getOrderId() : null,
-                country: isset($metric['country_id']) ? $countryMap[$metric['country_id']]->getCode() : null,
-                device: isset($metric['device_id']) ? $deviceMap[$metric['device_id']]->getType() : null,
-            );
-            $metricMap[$metricKey] = (int)$metric['id'];
-        }
-
-        return $metricMap;
-    }
-
-    /**
-     * @throws \Doctrine\DBAL\Exception
-     * @throws Exception
-     */
-    protected static function getChanneledMetricMap(
-        EntityManager $manager,
-        string $sql,
-        array $params,
-        array $metricMap,
-    ): array {
-        // Update the metric map with the newly inserted channeled metrics
-        $existingChanneledMetrics = $manager->getConnection()
-            ->executeQuery($sql, $params)
-            ->fetchAllAssociative();
-
-        // Map channeled metrics to their IDs and create a map for quick access
-        $channeledMetricMap = [];
-        foreach ($existingChanneledMetrics as $channeledMetric) {
-            if (!isset($metricMap['mapReverse'][$channeledMetric['metric_id']])) {
-                throw new RuntimeException("Channeled metric with ID {$channeledMetric['id']} references non-existent metric ID {$channeledMetric['metric_id']}");
-            }
-            $metricKey = KeyGenerator::generateChanneledMetricKey(
-                channel: $channeledMetric['channel'],
-                platformId: $channeledMetric['platformId'],
-                metric: $channeledMetric['metric_id'],
-                platformCreatedAt: (new DateTimeImmutable($channeledMetric['platformCreatedAt']))->format('Y-m-d'),
-            );
-            $channeledMetricMap[$metricKey] = [
-                'id' => (int)$channeledMetric['id'],
-                'data' => $channeledMetric['data'],
-            ];
-        }
-
-        return $channeledMetricMap;
-    }
-
-    /**
-     * @throws \Doctrine\DBAL\Exception
-     */
-    protected static function getDimensionMap(
-        EntityManager $manager,
-        string $sql,
-        array $params
-    ): array {
-        // Update the metric map with the newly inserted dimensions
-        $existingDimensions = $manager->getConnection()
-            ->executeQuery($sql, $params)
-            ->fetchAllAssociative();
-
-        // Map dimensions to their IDs and create a map for quick access
-        $dimensionMap = [];
-        foreach ($existingDimensions as $dimension) {
-            $dimensionKey = KeyGenerator::generateChanneledMetricDimensionKey(
-                channeledMetric: $dimension['channeledMetric_id'],
-                dimensionKey: $dimension['dimensionKey'],
-                dimensionValue: $dimension['dimensionValue'],
-            );
-            $dimensionMap[$dimensionKey] = (int)$dimension['id'];
-        }
-
-        return $dimensionMap;
     }
 }
